TASK: Add AbstractUserAwareActionController
The AbstractUserAwareActionController holds information about the
currently logged in user and assigns the user object to the view. All
controllers which need information about the current user (have the
navbar partial in their template) should extend this controller.

closes #84
