+++ Classes/SquadIT/WebApp/Controller/StandardController.php
@@ -6,45 +6,9 @@ namespace SquadIT\WebApp\Controller;
  */
 
 use TYPO3\Flow\Annotations as Flow;
-use SquadIT\WebApp\Domain\Model\User;
-use SquadIT\WebApp\Domain\Repository\UserRepository;
-use TYPO3\Flow\Security\Context;
 
-class StandardController extends \TYPO3\Flow\Mvc\Controller\ActionController
+class StandardController extends AbstractUserAwareActionController
 {
-
-    /**
-     * @Flow\Inject
-     * @var User
-     */
-    protected $user;
-
-    /**
-     * @Flow\Inject
-     * @var UserRepository
-     */
-    protected $userRepository;
-
-    /**
-     * @Flow\Inject
-     * @var Context
-     */
-    protected $securityContext;
-
-    /**
-     * @param \TYPO3\Flow\Mvc\View\ViewInterface $view
-     * @return void
-     */
-    public function initializeView(\TYPO3\Flow\Mvc\View\ViewInterface $view)
-    {
-        $this->user = $this->userRepository->findOneByAccount($this->securityContext->getAccount());
-        if ($this->user === null) {
-            return;
-        }
-        $view->assign('user_firstname', $this->user->getFirstname());
-        $view->assign('user_initials', substr($this->user->getFirstname(), 0, 1) . substr($this->user->getLastname(), 0, 1));
-    }
-
     /**
      * @return void
      */
