+++ Classes/SquadIT/WebApp/Controller/UserController.php
@@ -8,32 +8,18 @@ namespace SquadIT\WebApp\Controller;
 use TYPO3\Flow\Annotations as Flow;
 use TYPO3\Flow\Error\Message;
 use TYPO3\Flow\Security\Account;
-use TYPO3\Flow\Security\AccountRepository;
 use TYPO3\Flow\Security\AccountFactory;
 use TYPO3\Flow\Security\Cryptography\HashService;
 use SquadIT\WebApp\Domain\Model\User;
-use SquadIT\WebApp\Domain\Repository\UserRepository;
 
-class UserController extends \TYPO3\Flow\Mvc\Controller\ActionController
+class UserController extends AbstractUserAwareActionController
 {
-
-    /**
-     * @Flow\Inject
-     * @var AccountRepository
-     */
-    protected $accountRepository;
-
     /**
      * @Flow\Inject
      * @var AccountFactory
      */
     protected $accountFactory;
 
-    /**
-     * @Flow\Inject
-     * @var \TYPO3\Flow\Security\Context
-     */
-    protected $securityContext;
 
     /**
      * @Flow\Inject
@@ -42,36 +28,10 @@ class UserController extends \TYPO3\Flow\Mvc\Controller\ActionController
     protected $hashService;
 
     /**
-     * @Flow\Inject
-     * @var UserRepository
-     */
-    protected $userRepository;
-
-    /**
-     * @var User
-     */
-    protected $user;
-
-    /**
-     * @param \TYPO3\Flow\Mvc\View\ViewInterface $view
-     * @return void
-     */
-    public function initializeView(\TYPO3\Flow\Mvc\View\ViewInterface $view)
-    {
-        $this->user = $this->userRepository->findOneByAccount($this->securityContext->getAccount());
-        if ($this->user === null) {
-            return;
-        }
-        $view->assign('user_firstname', $this->user->getFirstname());
-        $view->assign('user_initials', substr($this->user->getFirstname(), 0, 1) . substr($this->user->getLastname(), 0, 1));
-    }
-
-    /**
      * @return void
      */
     public function indexAction()
     {
-        $this->view->assign('user', $this->user);
         $this->view->assign('account', $this->securityContext->getAccount());
     }
 
@@ -135,7 +95,7 @@ class UserController extends \TYPO3\Flow\Mvc\Controller\ActionController
      * @param User $user
      * @return void
      */
-    public function updateAction(User $user)
+    public function updateAction($user)
     {
         $this->userRepository->update($user);
         $this->addFlashMessage('Updated your profile');
