+++ Resources/Private/Partials/Navbar.html
@@ -33,8 +33,8 @@
       <ul class="nav navbar-nav navbar-right">
         <li class="dropdown">
           <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">
-              <img class="img-circle" src="//place-hold.it/30/888/fff&text={user_initials}&fontsize=14" alt="Profile Picture" />
-              {user_firstname} <span class="caret"></span>
+              <img class="img-circle" src="//place-hold.it/30/888/fff&text={user.firstname -> f:format.crop(maxCharacters: 1, append: '')}{user.lastname -> f:format.crop(maxCharacters: 1, append: '')}&fontsize=14" alt="Profile Picture" />
+              {user.firstname} <span class="caret"></span>
           </a>
           <ul class="dropdown-menu">
             <li><f:link.action controller="user" action="index">My Profile</f:link.action></li>
